move project from within other project structure
