+++ ConfigTransformation/packages.config
@@ -6,10 +6,10 @@
   <package id="EnvDTE90" version="9.0.1" targetFramework="net45" />
   <package id="EnvDTE90a" version="9.0.1" targetFramework="net45" />
   <package id="Marazt.Commons.Xml" version="1.0.14270.1006" targetFramework="net45" />
-  <package id="Microsoft.VisualStudio.CoreUtility" version="15.0.26201" targetFramework="net462" />
+  <package id="Microsoft.VisualStudio.CoreUtility" version="15.0.26606" targetFramework="net462" />
   <package id="Microsoft.VisualStudio.OLE.Interop" version="7.10.6070" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Shell.Framework" version="15.0.26201" targetFramework="net462" />
-  <package id="Microsoft.VisualStudio.Utilities" version="15.0.26201" targetFramework="net462" />
+  <package id="Microsoft.VisualStudio.Shell.Framework" version="15.0.26606" targetFramework="net462" />
+  <package id="Microsoft.VisualStudio.Utilities" version="15.0.26607" targetFramework="net462" />
   <package id="Microsoft.Web.Xdt" version="2.1.1" targetFramework="net45" />
   <package id="stdole" version="7.0.3301" targetFramework="net45" />
 </packages>
