+++ ConfigTransformation_IntegrationTests/Key.snk
